Disable attempts to auto-disconnect voice
Hopefully this fixes problems with the responsiveness of the stop_voice
command.

Seems like the bot's loop got hung up on... not?... doing end_playback?
I dunno, but after= has not been working for me so it only makes sense
to not use it until I can get it working.