More extra setup info
After fiddling more with setting up a Raspberry Pi for a server, I
ended up with some convenience things that deserved documenting.

The instructions are still a mess, INSTALL is still WIP. This is in
part for my own benefit, since frankly CrabBot is still not finished
enough for finalized instructions.