Change the audio player task to be ended and restarted as needed
Hopefully this should make it less likely to lock up mysteriously. Or
at least this should cleanup the bot loop task list so it's not always
waiting around for audio to play.

There's also more reliance on the player task to clean itself up, to
reduce the amount of voice disconnection points and related separate
logging strings.