Minor changes to work on Windows
I should've just blocked out the readline import from the start, since it isn't that important, and it's possible for readline to not work on systems other than Windows.

But I did not expect Windows Python to get choked up by a text file with newer emoji in it (cakes.txt, the fish cake emoji) because Windows "prefers" some other encoding. Kinda mad about that, honestly. I fully expect the explicit encoding change to invert the problem somehow and hoist me in the future.