Move Messages and Quotes cogs to be loaded by CrabBot instead of main
Mostly this is just to see how difficult it would be. I don't remember why I didn't just do this before.

I technically like keeping CrabBot itself kind of generic, but I also find that having __main__ add CrabBot's commands feels kind of janky/confusing.

Tacking on cog-specific args to CrabBot also feels more messy than the __main__ version, but it does feel like it's all in a more relevant spot. I suppose it depends on whether you think of CrabBot as an interface to run cogs, or as a whole bot that is implemented with cogs. I'm not sure which I lean towards, so the non-commital compromise it to call some cogs "defaults".

I do however like having the Voice cog stay in __main__, both as a usage example and as a way to keep the mess around Voice (eg. the code/argparses to keep it optional) out of CrabBot.