Initial basic poetry support
Seems to work alright for dependency management without too much fuss.

This almost certainly should not be used for packaging yet (if at all). At least, I'd like to move __main__.py into the actual package and rework it for running as a package, but that's going to be disruptive, so I haven't done it yet.