Move __main__.py to package, and document resulting new package usage
I think it was mostly leftover that I had it sitting outside to run the package, even though it was named correctly to be used with -m to run the package automagically.

Someone tell my younger self that `python -m crabbot` looks more cool and professional, heh.

(I was gonna change __main__.py to use relative imports, but the sys.modules usage complicates it so it might as well need absolute imports anyway.)