Adjust default paths to support packaging, adjust .gitignore accordingly
OK, as far as I've seen this is way easier than I figured, and using it as a package is super convenient.

Also changed the default for quotes to be the current directory to be less mysterious, even though the previous idea was to not pollute the source directory. Hopefully now that Poetry makes running in different directories less of a pain, it won't even be an issue, but I'm updating the .gitignore to try to ensure it doesn't happen in the first place.